backend/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-cors==3.0.10
flask-jwt-extended==4.4.4
geopy==2.3.0
python-dotenv==1.0.0
Pillow==10.0.0